Modifica serializer de aluno e insere informaçoes de nome, escola e dre na criaçao do aluno
